Style main page to use box model properties
Per assignment instructions in the Block and
Inline lesson on TOP I have reformatted the main
page to use box model properties to better style
the page.

I had to hardcode a value for the reciple list so
that the background colour for its parent ul
element can take the full page and not force
scrolling (at least on my machine).

This is definitely an issue that I hope to resolve
in the future.

I'm also playing with better commit messages. No
idea if I'm doing this right...